requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
pyyaml==6.0.1
webdriver-manager==4.0.1
loguru==0.7.2
python-dotenv==1.0.0
cryptography==41.0.5 